+++ undo/src/stream/mod.rs
@@ -1,12 +1,10 @@
 mod handlers;
-mod process;
 mod topics;
 mod undo;
 mod xadd;
 mod xread;
 
 pub use handlers::init;
-pub use process::*;
 pub use xadd::*;
 pub use xread::*;
 
@@ -14,7 +12,6 @@ use bot_model::api::*;
 use move_model::GameState;
 use undo_model::api::*;
 
-pub const GROUP_NAME: &str = "undo";
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum StreamOutput {
